+++ hydra-node/exe/tx-cost/Main.hs
@@ -3,6 +3,7 @@ import Hydra.Prelude hiding (catch)
 import Data.ByteString (hPut)
 import Data.Fixed (Centi)
 import Hydra.Cardano.Api (Lovelace (Lovelace))
+import Hydra.Contract (ScriptInfo (..), scriptInfo)
 import Hydra.Ledger.Cardano.Evaluate (maxCpu, maxMem, maxTxSize)
 import Options.Applicative (
   Parser,
@@ -83,6 +84,7 @@ writeTransactionCostMarkdown hdl = do
     encodeUtf8 $
       unlines $
         pageHeader
+          <> scriptSizes
           <> intersperse
             ""
             [ initC
@@ -119,6 +121,26 @@ pageHeader =
 {-# NOINLINE now #-}
 now :: UTCTime
 now = unsafePerformIO getCurrentTime
+
+scriptSizes :: [Text]
+scriptSizes =
+  [ "## Script summary"
+  , ""
+  , "| Name   | Hash | Size (Bytes) "
+  , "| :----- | ---: | -----------: "
+  , "| " <> "νInitial" <> " | " <> show initialScriptHash <> " | " <> show initialScriptSize <> " | "
+  , "| " <> "νCommit" <> " | " <> show commitScriptHash <> " | " <> show commitScriptSize <> " | "
+  , "| " <> "νHead" <> " | " <> show headScriptHash <> " | " <> show headScriptSize <> " | "
+  ]
+ where
+  ScriptInfo
+    { initialScriptHash
+    , initialScriptSize
+    , commitScriptHash
+    , commitScriptSize
+    , headScriptHash
+    , headScriptSize
+    } = scriptInfo
 
 costOfInit :: IO Text
 costOfInit = markdownInitCost <$> computeInitCost